requirements.txt
Django==5.0.6
djangorestframework==3.15.2
psycopg[binary]==3.2.1
python-dotenv==1.0.1
pytest==8.3.2
pytest-django==4.8.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
locust==2.31.6
drf-spectacular==0.27.2 
drf-spectacular-sidecar==2024.7.1